requirements.txt
torch>=2.0.1
torchvision>=0.15.2
timm>=0.9.2
fastapi>=0.101.1
uvicorn>=0.23.2
python-multipart>=0.0.6
matplotlib>=3.7.2
seaborn>=0.12.2
numpy>=1.25.2
pillow>=10.0.0
scikit-learn>=1.3.0
grad-cam>=1.4.8
